app/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.8
distlib==0.3.6
filelock==3.12.2
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.216
langchainplus-sdk==0.0.17
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pip==23.1.2
platformdirs==3.8.0
pydantic==1.10.9
pyee==11.0.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
setuptools==67.8.0
six==1.16.0
slack==0.0.2
slack-bolt==1.18.0
slack-sdk==3.21.3
slackclient==2.9.4
slackeventsapi==3.0.1
SQLAlchemy==2.0.17
tenacity==8.2.2
tqdm==4.65.0
typing_extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
Werkzeug==2.3.6
wheel==0.38.4
yarl==1.9.2
cryptography>=3.4.7